vectorizers set types to int32

+++ python/baseline/vectorizers.py
@@ -195,7 +195,7 @@ class Char2DVectorizer(AbstractCharVectorizer):
 
         EOW = vocab.get('<EOW>', vocab.get(' '))
 
-        vec2d = np.zeros((self.mxlen, self.mxwlen), dtype=int)
+        vec2d = np.zeros((self.mxlen, self.mxwlen), dtype=np.int32)
         i = 0
         j = 0
         for atom in self._next_element(tokens, vocab):
@@ -258,7 +258,7 @@ class Char1DVectorizer(AbstractCharVectorizer):
         if self.mxlen < 0:
             self.mxlen = self.max_seen_tok
 
-        vec1d = np.zeros(self.mxlen, dtype=int)
+        vec1d = np.zeros(self.mxlen, dtype=np.int32)
         for i, atom in enumerate(self._next_element(tokens, vocab)):
             if i == self.mxlen:
                 i -= 1
